refactor backend into smaller modules and setup new router for user requests
